==> 1_sentiment/1.0_sample_sentiment_sentences.ipynb <==
# Randomly sample sentences from natural and synthetic corpora with which to compute Sentiment 

Authors: Naomi Baes and Chat GPT

Input:
- Natural corpus: 0.0_corpus_preprocessing/output/natural_lines_targets/"abuse_lines_targets" etc.
- Synthetic corpus: 
    1_sentiment/synthetic/output/5-year/"abuse_1970-1974.synthetic_sentences.csv" etc.
    1_sentiment/synthetic/output/all-year/"abuse_synthetic_sentences.csv" etc.

Output:
- 5-year sampling strategy
    5-year/negative/"abuse_1970-1974_synthetic_sentiment_0.1.tsv" etc.
    5-year/positive/"abuse_1970-1974_synthetic_sentiment_0.1.tsv" etc.
- all-year sampling strategy
    5-year/negative/"abuse_synthetic_sentiment_0.1.tsv" etc.
    5-year/positive/"abuse_synthetic_sentiment_0.1.tsv" etc.

# Randomly sample using the random stratified sampling strategy (5-year intervals)
%run step0_randomly_sample_sentences_5-year.py

# Randomly sample using the bootstrapped sampling strategy (whole corpus)
%run step0_randomly_sample_sentences_all-year.py
