==> 1_sentiment/1.1_sentiment-index-absa.ipynb <==
# Compute Sentiment Index (ABSA)

Author: Raphael Merx

Original: https://huggingface.co/MoritzLaurer/deberta-v3-large-zeroshot-v2.0
ABSA variant: https://huggingface.co/yangheng/deberta-v3-base-absa-v1.1

Input: raw sentences

## 5-year Sentiment Index

%run step1_absa_sentiment_5-year_with_SE.py

Get standard errors of the mean for the sentiment indices

## all-year Sentiment Index

%run step1_absa_sentiment_all-year_with_SE.py # 1 hour with GPU; 304m 44.6s hours with CPU

# End of script
